criei componentes que facilitam a adição de cursos e aulas para o professor, plus ProfessorPage
