+++ src/index.js
@@ -10,17 +10,24 @@ import LoginPage from './pages/LoginPage';
 import CatalogoCursos from './pages/CatalogoCursos';
 import Teste from './pages/Teste';
 import Cursos from './pages/Cursos/Cursos';
+import ProfessorPage from './pages/professorPage/ProfessorPage';
+import DefaultTemplate from './Componentes/Templates/DefaultTemplate';
+import ProfessorAddAula from './pages/ProfessorAddAula/ProfessorAddAula';
 
 ReactDOM.render(
   <BrowserRouter>
     <Routes>
       <Route path="/" element={<App />}>
-        <Route path="/" element={<Main />} />
         <Route path="cadastro" element={<CadastroPage />} />
         <Route path="login" element={<LoginPage />} />
-        <Route path="CatalagodeCursos" element={<CatalogoCursos />} />
         <Route path="testes" element={<Teste />}/>
-        <Route path="cursos" element={<Cursos />} />
+        <Route element={<DefaultTemplate />} >
+          <Route path="/" element={<Main />} /> 
+          <Route path="cursos" element={<Cursos />} />
+          <Route path="CatalagodeCursos" element={<CatalogoCursos />} />  
+          <Route path="professors" element={<ProfessorPage />} />
+          <Route path="ProfessorAddAula" element={<ProfessorAddAula />} />
+        </Route>
       </Route>
     </Routes>
   </BrowserRouter>,
